Revert "Docs add readme"
