+++ README.md
@@ -1,22 +1,8 @@
-# Getting Started
+This is a [Next.js](https://nextjs.org/) project bootstrapped with [`create-next-app`](https://github.com/vercel/next.js/tree/canary/packages/create-next-app).
 
-## Usage
+## Getting Started
 
-1. Fork this repository.
-2. Sign up for [Alchemy](https://www.alchemy.com/) and [Ark Project](https://www.arkproject.dev/). Obtain your API keys.
-3. Create a `.env.local` file and populate it with the environment variables as specified in the `.env.example` file.
-
-### installation
-
-```bash
-npm install
-#or
-yarn
-#or
-pnpm install
-```
-
-### Run development server
+First, run the development server:
 
 ```bash
 npm run dev
@@ -24,20 +10,27 @@ npm run dev
 yarn dev
 # or
 pnpm dev
+# or
+bun dev
 ```
 
-## Environment variables
+Open [http://localhost:3000](http://localhost:3000) with your browser to see the result.
 
-We use Ark Project for indexing NFT information and Alchemy for blockchain interactions. Ensure you have accounts with both platforms and input your API keys in the `.env.local` file.
+You can start editing the page by modifying `app/page.tsx`. The page auto-updates as you edit the file.
 
-## Getting your NFT artwork to show up in the iFrame
-
-- NFT Route: The contract address and token ID are extracted from the URL parameters `/contractAddress/tokenId`. The `fetchNFTData` function in `app/hooks/index.ts` fetches the NFT data, and the Token Bound Account (TBA) associated with the token ID (used as the salt) is returned.
-
-- TBA Route: The contract address of the TBA is retrieved from the URL parameters `/tokenboundAddress`. The getOwnerNFT function in `app/utils/index.ts` retrieves the owner NFT’s contract address and token ID. The NFT data is then fetched using the fetchNFTData function, also located in `app/utils/index.ts`.
+This project uses [`next/font`](https://nextjs.org/docs/basic-features/font-optimization) to automatically optimize and load Inter, a custom Google Font.
 
 ## Learn More
 
-To learn more about Tokenbound, take a look at the following resources:
+To learn more about Next.js, take a look at the following resources:
 
-- [Starknet Tokenbound Documentation](https://docs.tbaexplorer.com/)
+- [Next.js Documentation](https://nextjs.org/docs) - learn about Next.js features and API.
+- [Learn Next.js](https://nextjs.org/learn) - an interactive Next.js tutorial.
+
+You can check out [the Next.js GitHub repository](https://github.com/vercel/next.js/) - your feedback and contributions are welcome!
+
+## Deploy on Vercel
+
+The easiest way to deploy your Next.js app is to use the [Vercel Platform](https://vercel.com/new?utm_medium=default-template&filter=next.js&utm_source=create-next-app&utm_campaign=create-next-app-readme) from the creators of Next.js.
+
+Check out our [Next.js deployment documentation](https://nextjs.org/docs/deployment) for more details.
