+++ app/tba/[tokenboundAddress]/[chainId]/page.tsx
@@ -1,13 +1,14 @@
 "use client";
 import Panel from "@/app/components/Panel";
+import { getOwnerNFT } from "@/hooks";
+import { TBALogo2 } from "@/public/svg/Icons";
 import {
   NetworkType,
+  fetchNFTData,
   fetchTbaFungibleAssets,
   fetchTbaNonFungibleAssets,
   getChainData,
-} from "@/app/helper";
-import { getOwnerNFT, fetchNFTData } from "@/app/hooks";
-import { TBALogo2 } from "@/public/svg/Icons";
+} from "@/utils";
 import { useParams } from "next/navigation";
 import { useEffect, useState } from "react";
 import { num } from "starknet";
@@ -67,6 +68,9 @@ const TokenBound = () => {
       });
     };
     if (chainId && tokenboundAddress) {
+      // const { network, url } = getChainData(chainId.toUpperCase());
+      console.log(network);
+
       fetchOwnerNFT({ network, url });
       fetchTbaNonFungibleAssets({
         address: tokenboundAddress,
